Fix odd scrolling bug
Reason for bug: footer (which is invisible) was showing on screen. As
new content was inserted into the list, the RecyclerView made sure
that the footer was still on screen in an attempt to not disrupt user
reading. Naturally this made no sense since it's a spacing footer.

Solution: report the list as empty when it's actually empty, and so
ALL items (including footer) are inserted as new.

Also change the onInsert and onRemove callbacks to take the footer
into account. Going from Empty or to Empty state will call
notifyDataSetChanged instead of onInsert/onRemove.