Changed so that an empty feed can be dragged to be refreshed
The drag to refresh pattern is so convenient that it has become a
major annoyance that you can't use the same action to refresh an empty
feed. First you'd need to show all items so that the list was
populated and then you'd be able to drag to refresh. But then you'd be
forced to hide the read items again.

This changes to that the list is not actually hidden which allows the
drag to refresh action to be performed always. There is no difference
to the user since the list is empty and thus "invisible" anyway.

Fixes #40