Fixed concurrency of sync
Parallelizes the parsing of feeds, but not items.

Also adds a test for the synchronization.

The problem I was seeing was in fact the database transaction acting
up. When running concurrent operations inside the same transaction all
hell broke loose and the database transaction actually hanged the
entire flow.

Removed the transaction, it was something of a premature optimization
anyway, and opted for concurrent parsing instead.