Adjusted sync logic
The last_sync field is always set, even if the content didn't
change. This allows it to be used as a flag for when we last fetched
it and we never try to fetch a feed again within 15mins unless the
user manually sync it.

If more than 15mins has passed, then an attempt to fetch the feed is
made. If a cached or not-changed response is retrieved, then no feed
parsing is performed as before.

This should mean even fewer http requests as we before kept trying
feeds everytime even if they had been tried recently and they rarely
changed.