Added textDirection 'anyRtl' to all TextViews
The default detection goes by the first character, and if that happens
to be a Latin word then it will format the paragraph as LTR, even if
the rest of the paragraph is actually RTL the first word is just an
exception.

Instead, anyRtl will format the paragraph as RTL if it contains ANY
RTL markers.

Fixes #176