Added EphemeralState ViewModel to keep track of list scroll state
Refactoring to Navigation Component had a side effect of killing the
state of the list when you read an item (because nav component doesn't
keep a fragment back stack, it replaces the fragment always).

This introduces a ViewModel which has data bound to the activity
lifecycle, essentially mimicking onSaveInstance logic for the nav component.

Fixes #219