lang: updated Turkish Translation (#658)
* Update Turkish Translation

I re-translated the file by syncing it with the English string file, so it might look complicated. While there are 2 spaces on the left side of the English strings in the translations, the existing Turkish file has 4 spaces. I left it as it should be. Since I’m not using Weblate, I’m opening a PR; if you don’t want to merge it, I’d appreciate it if you could at least upload it to Weblate.

* Update strings.xml