feat: Android 10 (sdk29) is now the oldest supported version
Removing Conscrypt means relying on platform HTTPS support and TLS1.3 was
enabled in SDK29 (Android 10). Hence bumping to that.

This was released in 2019. So in effect supporting devices which received
their last OS update 6 years ago.

In Google play, Jun 2nd 2025, 1.82% of users are on a version of Android
older than 10.

Signed-off-by: Jonas Kalderstam <jonas@cowboyprogrammer.org>